@@ -1,57 +1,57 @@
 export interface UserLogin {
-  email: string;
-  password: string;
+  email: string
+  password: string
 }
 
 export interface User {
-  id: number;
-  displayName: string;
-  email: string;
-  password: string;
-  role: string;
+  id: number
+  displayName: string
+  email: string
+  password: string
+  role: string
 }
 
 export interface EditUser {
-  id: number;
-  name: string;
-  orientador: number;
-  email: string;
-  role: string;
+  id?: number | undefined
+  name?: string | undefined
+  orientador?: number | undefined
+  email?: string | undefined
+  role?: string | undefined
 }
 
 export interface EditFinance {
-  value: string;
-  type: number;
-  description: string;
-  user: User;
+  value: string
+  type: number
+  description: string
+  user: User
 }
 
 export interface AuthenticationResponse {
-  token: string;
-  user: AuthenticatedUser;
+  token: string
+  user: AuthenticatedUser
 }
 
 export interface AuthenticatedUser {
-  id: number;
-  displayName: string;
-  email: string;
-  role: string;
+  id: number
+  displayName: string
+  email: string
+  role: string
 }
 
 export interface Authorities {
-  authority: string;
+  authority: string
 }
 
 export interface Project {
-  id: number;
-  description: string;
-  subject: string;
+  id: number
+  description: string
+  subject: string
 }
 
 export interface Teacher {
-  id: number;
-  name: string;
-  email: string;
+  id: number
+  name: string
+  email: string
 }
 
 export type SignUpParams = {
@@ -62,38 +62,38 @@ export type SignUpParams = {
 }
 
 export interface SolicitationAudit {
-  id: number;
-  changeDate: string;
-  newStatus: string;
-  dropdown: boolean;
+  id: number
+  changeDate: string
+  newStatus: string
+  dropdown: boolean
   solicitation: {
-    changeDate: string;
-    solicitation: any;
-    id: number;
-    newStatus: string;
+    changeDate: string
+    solicitation: any
+    id: number
+    newStatus: string
     createdBy: {
-      name: string;
+      name: string
     }
     equipment: {
-      form: string;
-      name: string;
-    };
+      form: string
+      name: string
+    }
     project: {
-      description: string;
-      subject: string;
+      description: string
+      subject: string
       teacher: {
-        name: string;
+        name: string
       }
     }
-    value: number;
-    fileUrl: string;
+    value: number
+    fileUrl: string
     projectNature: {
-      content: string;
+      content: string
     }
   }
 }
 
 export type LabelValue = {
-  label: string,
+  label: string
   value: string
 }
