@@ -1,4 +1,4 @@
 export interface Partner {
-  id?: number
-  name: string
+  id?: number;
+  name: string;
 }
